pyproject.toml
[tool.poetry]
name = "cp-modeling"
version = "0.1.0"
description = "Computational Pathology model training"
authors = ["vladpopovici <popovici@bioxlab.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
neuraxle-tensorflow = "^0.1.2"
tensorflow = "^2.8.0"
autokeras = "^1.0.18"
Pillow = "^9.0.1"
numpy = "^1.22.2"
albumentations = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
